Refactoring and update dependencies
