+++ manifest.json
@@ -2,7 +2,7 @@
   "name": "@deskpro-apps/teamviewer",
   "title": "TeamViewer",
   "description": "Create TeamViewer sessions and share them with your users in ticket.",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "scope": "agent",
   "isSingleInstall": false,
   "hasDevMode": true,
