+++ package.json
@@ -13,8 +13,9 @@
     "test:coverage": "cross-env NODE_OPTIONS=--max-old-space-size=1024 jest --maxWorkers=75% --coverage"
   },
   "dependencies": {
-    "@deskpro/app-sdk": "1.1.9",
-    "@deskpro/deskpro-ui": "2.9.1",
+    "@deskpro/app-sdk": "1.2.8",
+    "@deskpro/deskpro-ui": "4.7.0",
+    "@heroicons/react": "1.0.6",
     "date-fns": "^2.29.3",
     "iframe-resizer": "^4.3.2",
     "lodash": "^4.17.21",
@@ -24,7 +25,7 @@
     "react-dom": "^17.0.2",
     "react-error-boundary": "^3.1.4",
     "react-router-dom": "^6.4.0",
-    "styled-components": "^5.3.5",
+    "styled-components": "^5.3.6",
     "ts-pattern": "4.0.5",
     "uuid": "^9.0.0"
   },
