+++ src/components/ActiveSessions/index.ts
@@ -1,0 +1,1 @@
+export { ActiveSessions } from "./ActiveSessions";
