+++ src/components/ExpiredSessions/index.ts
@@ -1,0 +1,1 @@
+export { ExpiredSessions } from "./ExpiredSessions";
