+++ src/components/SessionTitle/index.ts
@@ -1,0 +1,1 @@
+export { SessionTitle } from "./SessionTitle";
